추가: 드래그시 props의 onDrag 콜백으로 다시 부모로 돌려주기 - MouseUp시 draggedId = null이 되도록 설정

+++ todo-1/src/components/TodoItem.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { Item } from "../types";
 interface TodoItemProp {
   item: Item;
-  onDelete(id: number);
+  onDelete?: (id: number) => void;
+  onDrag?: (id: number) => void;
 }
 function getPrettyDate(date: Date) {
   const now = +new Date();
@@ -19,14 +20,21 @@ export class TodoItem extends React.Component<TodoItemProp> {
   constructor(props) {
     super(props);
     this.handleDelete = this.handleDelete.bind(this);
+    this.handleDrag = this.handleDrag.bind(this);
   }
   handleDelete() {
+    if (!this.props.onDelete) return;
     this.props.onDelete(this.props.item.id);
+  }
+  handleDrag(event) {
+    if (!this.props.onDrag) return;
+    if (event.target.closest("button")) return;
+    this.props.onDrag(this.props.item.id);
   }
   render() {
     const { item } = this.props;
     return (
-      <div className="todo-item">
+      <div onMouseDown={this.handleDrag} className="todo-item">
         <div>{item.content}</div>
         <div className="item-added-at">{getPrettyDate(item.addedAt)}</div>
         <button onClick={this.handleDelete}>X</button>
